+++ src/main/java/org/example/back_end_labs/controller/UserController.java
@@ -1,9 +1,12 @@
 package org.example.back_end_labs.controller;
 
+import jakarta.validation.constraints.NotBlank;
 import org.example.back_end_labs.model.User;
 import org.example.back_end_labs.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -11,6 +14,7 @@ import java.util.Optional;
 
 @RestController
 @RequestMapping("/")
+@Validated
 public class UserController {
     private final UserService userService;
 
@@ -20,25 +24,39 @@ public class UserController {
     }
 
     @PostMapping("/user")
-    public User createUser(@RequestParam String name) {
-        return userService.addUser(name);
+    public ResponseEntity<?> createUser(@RequestParam @NotBlank(message = "User name cannot be empty")
+                                            String name) {
+        User createdUser = userService.addUser(name);
+        return ResponseEntity.status(HttpStatus.CREATED).body(createdUser);
     }
 
     @GetMapping("/user/{userId}")
-    public ResponseEntity<User> getUserById(@PathVariable Long userId) {
-        return userService.getUserById(userId)
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
+    public ResponseEntity<?> getUserById(@PathVariable Long userId) {
+        Optional<User> user = userService.getUserById(userId);
+        if (user.isPresent()) {
+            return ResponseEntity.ok(user.get());
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User with ID " + userId + " not found.");
+        }
     }
 
     @DeleteMapping("/user/{userId}")
-    public ResponseEntity<Void> deleteUser(@PathVariable Long userId) {
-        userService.deleteUser(userId);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<String> deleteUser(@PathVariable Long userId) {
+        Optional<User> user = userService.getUserById(userId);
+        if (user.isPresent()) {
+            userService.deleteUser(userId);
+            return ResponseEntity.ok( "User with ID " + userId + " deleted successfully.");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User with ID " + userId + " not found.\nUser cannot be deleted.");
+        }
     }
 
     @GetMapping("/users")
-    public List<User> getAllUsers() {
-        return userService.getAllUsers();
+    public ResponseEntity<?> getAllUsers() {
+        List<User> users = userService.getAllUsers();
+        if (users.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("No users found.");
+        }
+        return ResponseEntity.ok(users);
     }
 }
